Add Jan's fixes from https://github.com/JvanKatwijk/qt-dab/commit/2e9f78574aa33e4b65385eda79d8a8ea743ab7a9

+++ src/backend/msc-handler.cpp
@@ -90,7 +90,7 @@ void mscHandler::stop_service(DescriptorType * d, int flag)
 void mscHandler::stop_service(int subchId, int flag)
 {
   locker.lock();
-  for (size_t i = 0; i < theBackends.size(); i++)
+  for (int32_t i = 0; i < (int32_t)theBackends.size(); i++)
   {
     Backend * b = theBackends.at(i);
     if ((b->subChId == subchId) && (b->borf == flag))
